improve astar router
